[mojo-lang] Remove old hack that breaks generics/traits. (#37075)
This removes a hack introduced in #25867.  It appears that my change
last week to how Self is represented in traits defines the need for
this away, and the integration test continues to pass.  This was
blocking #37059 by causing vtable references to not fold and types
to be considered unequal.  That patch will test this.

MODULAR_ORIG_COMMIT_REV_ID: 46a5d838adacf779ce89ca1b3d0858a685341f66