[Kernels][GPU] Optimize double buffer gemm 1.1x using int32 indexing (#36992)
Analysis in

[Internal link]

This improves the double buffer gemm 12% on A100, from 16.4 TFLOPs to
18.1 TFLOPs (cublas ~18.4 TFLOPs) by reducing the max register usage.

Since the LayoutTensor is fully static we can use Int32 indexing when an
Int32 can fully address the size. This isn't a long term solution but it
works for now.

All of the functions in id.mojo should also return an Int32, which I'll
do in a followup.

LAYOUT_ORIG_REV_ID: 599da50a5141f3a16b7a5631e82b401d1b1778dc