[stdlib] Initial changes for AnyPointer to UnsafePointer stdlib migration (#37135)
This change migrates the uses of `AnyPointer` to `UnsafePointer` within
the Stdlib plus some other low-hanging fruit use cases. I didn't change
the `AnyPointer` implementation to `UnsafePointer` because it generated
compiler errors outside of the stdlib that I will resolve in a future
change. This change adds an alias for `UnsafePointer` that will be
removed in the future as it is only temporarily useful during the
migration.

This change is in support of #37054.

MOJO_MAX_ORIG_REV_ID: d72ab52e12df6d7c7ce90c61a9e06b9c20a5e507