[Kernels] Align packed B to be multiple of the kernel width (#37106)
With bfloat16 and AVX512 the tile_inner_size is 128 but N is 64. We pack
using N // tile_inner_size which which is zero. Instead we should pack
using align_up(N, tile_inner_size).

This was fixed in #34599 but for some reason this was removed.

LINALG_ORIG_REV_ID: 6373d794034dd0b61bdba5182858e7ee6408420a