[KGEN] Correct Upward Consume Set of Conditional Self Init Regions (#36360)
This PR adds error handling support for elif. Details below.

Consider
```
lit.func @__init__2[mut *"L1", mut *"L2", imm *"L3"](
 %self: !lit.ref<!HasMemFields, mut *"L1"> init_self,
 %b: !lit.ref<!S, imm *"L3"> borrow_in_mem,
 %s : i1,|, ?,
 %__error__: !lit.ref<!Error, mut *"L2"> byref_error) throws -> i1 {

 // Initialize raising self.
 %6 = lit.call @__init__1[mut *"L2", mut *"L1"](%__error__, %self) : !lit.signature<[2](?, "__error__": !lit.ref<!Error, mut *[0,0]> byref_error, "__result__": !lit.ref<!HasMemFields, mut *[0,1]> byref_result) throws -> i1>

 // Check Outcome of throwing function and destroy intialized values upon failure
 hlcf.elif {
   // BLOCK 1
   hlcf.elif.yield %6 : i1
 } then {
   // BLOCK 2
   lit.ownership.mark_initialized %__error__ : <!Error, mut *"L2">
   %8 = kgen.param.constant: i1 = <1>
   lit.error_return %8 : i1
 } {
  // BLOCK 3
   hlcf.elif.yield %s : i1
 } then {
   // BLOCK 4
   lit.ownership.mark_initialized %__error__ : <!Error, mut *"L2">
   %9 = kgen.param.constant: i1 = <1>
   lit.error_return %9 : i1
 } else {
   // BLOCK 0
   lit.ownership.mark_initialized %self : <!HasMemFields, mut *"L1">
   hlcf.yield
 }
 %7 = kgen.param.constant: i1 = <0>
 kgen.return %7 : i1
}
```
Without intervention, the current algorithm will do the following in the
unification step:
1. KGEN.return marks self as totally initialized [1][111]
2. Block 0, mark as initialized [1][000]
3. Block 4 marks self as not initialized: [1][001]
4. Block 3 examines the upward set of its successors, which is [1][001],
which results in [0][001] difference that needs to be applied to block
0. This is incorrect! What we really want is no destructor in block 0
since block 4 did not actually consume the object because the object was
not initialized.

The upward consume set of {0,4} should still be [1][001] because it
marks the self object as indestructible
The upward consume set for the sake of calculating destructors in {0,4}
should be [0][000] to prevent destructors from being emitted into block
4.

To address this, we want to reset the full object bit in the upward
consume set of the regions r in succ(p) for some node p where the
successors either initialize the error or initialize the success region.

MODULAR_ORIG_COMMIT_REV_ID: 030b56a5c1e13c6060d1c74c4adef33566acde65