[Kernels] Fix Multiple Reductions When Reducing in Outer Dimension (#37249)
The old outer reduction for the case of multiple reductions produced incorrect results. The main issue is accumulated results for the reduction is stored in a type like `var result = StaticTuple[SIMD[_type, width], num_reductions]`

Where each index in the tuple is the accumulation of a particular reduction. In the old implementation of outer reduction, we mistakenly unrolled and indexed into `result`. This produced a `SIMD[_type, width]` which appears to be splatted to a tuple of length `num_reductions`. Of course, this splatting is incorrect since we only use the result of the first reduction.

This is now fixed and the extraneous unrolling is removed.

Also adds a test doing reduction along an outer dimension.

ALGORITHM_ORIG_REV_ID: 6185494a72aecdd93b00b9780c77aa4819439726