[mojo] Refactor specializeModuleForPreElaborationLinking (#37077)
With the recent cleanups and consolidation of package compilation,
the structure of these methods needs to change. The current methods
over-expose too much information about package compilation and
structure, which leads to dependencies on this behavior and usages
in areas that shouldn't be using them. This PR restructures the
exposed aspect of this method to mostly be a wrapper around running
the library generation pipeline (which can soon be incorporated into
a better Mojo compiler driver), and hides all of the details specific
to packages.

This also adds a new parser entry point method for getting
a self contained module from a package, which is part of what MOGG
was using this method for (which is no longer guaranteed or exposed).
This is quite a bit cleaner/explicit about what the intentions are,
and also allows for a followup to remove the half-baked self-contained
nature of packages (which greatly speeds up individual package
compilation, drops package size by >10x, and more).

MODULAR_ORIG_COMMIT_REV_ID: 954006f136c02ed309296ef666073485e2bf8e84