[mojo] Track a packages dependencies in a new LinkDependencyArrayAttr (#37078)
This will allow for more explicitly specifying the other packages that a
package depends on. In a followup this will be used to greatly shrink the
IR contained within a package, using the dependency information to
ensure dependencies are pulled in on package use.

MODULAR_ORIG_COMMIT_REV_ID: 5105bb33b0a705258ecf373d54f2f4f3106b4c14