[bazel] Rework shared libraries (#37433)
cc_shared_library fully links the dependencies in the tree. This causes
issues for us since our globals libraries depends on a large portion of
the dependency tree, but don't use most of it in the library itself. The
result is that most of the symbols are dropped, but dynamic_deps makes
cc_binary avoid duplicating linking libraries, which results in linker
errors no matter what you do. In the most ideal case the dependency
trees of the globals libraries would be completely empty, and
cc_shared_library would be perfect. Instead we can use a more jank
version of this by using cc_binary(linkshared=True) which creates a
shared library with only the direct contents, leaving the rest up to the
final cc_binary to provide. I created a globals_library macro to
encapsulate the complexity of this approach.

MODULAR_ORIG_COMMIT_REV_ID: dcaab1e270dea6c379ad2efe1173b18bf619a9af