[mojo] Refactor post-parse package modules to not be self contained (#37079)
Packages are currently in a weird half state where they kind of contain
their dependencies, but also don't (and rely on the user of a package to
resolve dependencies at the same time). This not quite real dependency
management is a bit of a mess and leads to lots of duplicated IR inside
of a package (recursively stored dependent modules), and various other
weird situations. Until we actually work out how we want to model package
dependencies and resolution, unify around a state where the
package use needs to pull and resolve the dependencies. This leads to
a significantly simpler and smaller package (kernels-nn goes from 90mb
to 4mb).

MODULAR_ORIG_COMMIT_REV_ID: 82bb288cccb197ebcb6e0c740e7bc294ca422ebc