[bazel] Add opt_wrapper rule (#37574)
This forces a target to be built with optimizations, which is useful for
some lit tests where that makes a big difference. The downside of using
this is you split the caches between when you `./bazelw build KGEN:kgen`
and when you run tests that use this target. Unfortunately the
performance matters enough that we need it. We can add more to this
transition as needed. We'll probably want `-g` too so we can just
encourage the use of `./bazelw build KGEN:kgen.opt` when you're running
and testing so that you only build 1 thing. We _don't_ want to use
bazel's builtin exec transition here because of future remote execution
usage where we do need this binary to be built for the target machine
that it will run on, which might not be the host machine.

MODULAR_ORIG_COMMIT_REV_ID: 22c570c7b2fe2aa398ab28918d4562378f6df17a