[CI] Add a Green-Only Asan Workflow (#37242)
Adding a new ASAN workflow that runs specific targets that we know are
green as of now, until our "main" ASAN build is in a spot where running
everything is green. These will be bisected and run like ASAN, except
the notifications for failure will be sent to #ci-platform for higher
visibility (I am willing to keep sending them to ci-asan and just
monitor that too)

To accomplish this, I added a new semantic test tag "asan-red" that
should denote a test suite that fails when run on ASAN. Then I added a
new command to `build.py` that uses our (non-enforced) naming
conventions to tree search Ninja for all the test targets, then subtract
the tests that have been marked asan-red. It's kind of hacky, but it
allows us specify a CI run where we run "All post submit tests, minus
tests that we have marked asan-red"

The limitation to this is that crawling the build graph takes time (~30
seconds) with the current implementation, and we only mark large suites
as asan-red rather than individual tests. So this is a somewhat blunt
tool.

The idea is that now we have a green asan base we can work from, we will
strive to keep _this_ one green with bisecting.

MODULAR_ORIG_COMMIT_REV_ID: b5ac90de9612fce79501fc97f5ff210e277870fa