[mojo-tooling][mojo-stdlib] Add a hack to support the current stdlib OSS layout (#37616)
When we open sourced the standard library, we used a very forward looking directory
structure:

```
stdlib/
  ...
  src/
  ...
```

This is quite an ideal layout, but the problem is the fact that none of our current tooling
expects the package name (stdlib) to be different from the directory name (src). This ends
up breaking a huge number of the tools/infra that aren't expecting this (LSP, docs, etc.). Proper
support for this kind of layout is coming down the line (with the project spec work), but for
now this PR adds a hack in the tooling entry point to check whether the top-level package
is named "src" (in which case we use the parent directory name as the package name). This is
less than ideal, but will at least fix the development workflow of those using the Mojo stdlib
until we have proper support for this kind of thing.

Closes #36420
MODULAR_ORIG_COMMIT_REV_ID: 2acda8135f1dafa72d6594390483db2c76ee95b7