[mojo-lang] Teach parameter inference about implicit conversions. (#37625)
A long-standing weakness of our parameter inference system is that
it doesn't infer parameters when implicit conversions are required
to line things up.  This rectifies this, unblocking the address_of
changes that depend on it.

Note: I find it deeply unsatisfying that parameter inference has
to duplicate all the call resolution and value conversion logic
for parameters.  This whole thing needs a rethink!

modular-orig-commit: 299af9b663cc6e8aaae31eb11b73fd526557b8a0