[mojo-stdlib] Fix `address_of` member on pointer-like things. (#37629)
It isn't safe to take the address of an inout/borrowed argument that
might expand
to a register type, because those arguments will eventually be passed as
a register
(which doesn't have an address!) when IR generation is done.

One of the biggest violators of this has been the `Pointer.address_of`
sorts of
methods. Fix them to take a `Reference` instead of taking the address of
an inout
argument, which is also nice because they now don't require the target
to be
mutable (e.g. they'll work with borrowed memory arguments). This
requires
propagating the reference out through similar inout arguments and fixing
some
type inference holes.

This fixes #32654 and eliminates some last uses of
`__get_lvalue_as_address`,
which didn't do the proper checking for this problem.

EXTENSIBILITY_ORIG_REV_ID: e81cf8dd2e48bcd74bc6e86923fa8b0b5d536009