[mojo-lang] Improve parameter inference, resolving fixme. (#37664)
When we want to infer the second or later parameter for a function,
we already know about previous parameters.  The values of those
parameters can inform and specialize what is valid in subsequent
parameters, so we need to substitute them in before doing inference.
This eliminates a class of inference problems, and makes it easier
to work on this.

This patch resolves a FIXME I added in Jan 2023 in #7863 (!) which
first implemented parameter inference.

MODULAR_ORIG_COMMIT_REV_ID: 648b9b3fc58df944d06b04803f5c88aa16269949