[mojo-lang] Introduce a new `!kgen.pack.load` op. (#37673)
This introduces a new operation that can take a variadic pack of
pointers and return all the elements they hold.  This is lowered
after elaboration to a bunch of individual load operations. This
is needed to help get rid of the old VariadicPack path.

MODULAR_ORIG_COMMIT_REV_ID: 046aebcea232d6eaece65ccb008efa0ac22be475