[mojo-stdlib] Move gpu/host/function to fancy packs, NFC. (#37677)
The original code seems to work, but was mega-scary - it built
pointers to local variables declared on the stack, which would only
work because of the replication due to unroll.  This points to the
original memory in the pack, which aligns much better with what it
needs.  This also removes `AnyRegTuple`, huzzah.

This resolves the fourth item in #36916

GPU_ORIG_REV_ID: 83b8a386e9bc0d5c48516486436063cf07beb7c5