[mojo-stdlib] Rewrite ListLiteral to wrap `Tuple` (#37681)
This rewrites the `ListLiteral` type to wrap a `Tuple` instead of a
kgen.pack directly.
This enables it to have proper values semantics for the contained
operation, and work
with any `CollectionElement` type, instead of just `AnyRegType` types.
This also moves one random testcase off borrowed packs.

This is the last known thing using legacy-packs, so this fixes #36916!

modular-orig-commit: fcead2d1981eb29a46e53db9ac97e43be917de2b