[Kernels][GPU] Add a setup-gpu-benchmarking script (#37692)
This sets up the gpu-benchmarking script which is written in mojo using
the nvml library. This will lock the clock frequency on the gpu to the
max rate. It will also disable GPU turbo boost. Not all GPUs have the
ability to do this, so all this functionality is "best effort".

Future PRs will look at setting the persistence mode along with other
things.

Note that you need to run this command as sudo at this moment (will
write instructions on how to set it up so we can run without in a
followup PR).

Closes #24248

NVML_ORIG_REV_ID: c4a108600c828856a42b139338015bf5f72cd9ee