[bazel] Fix libKGENCompilerRT-static.a (#37703)
We distribute this library and link it into mojo binaries. Unlike other
globals libraries it never ends up as a dylib that eliminates what
doesn't need to be linked. This change includes a new rule that creates
a static library with only what is passed. This operation is inherently
unsafe, but is required in order to ensure that we produce the smallest
possible library since we are distributing it. If all transitives were
included the library would be 7gbs. Right now this produces a 80mb
library. This risk is reduced by the link test here that should validate
the required libraries are actually specified.

MODULAR_ORIG_COMMIT_REV_ID: d5705f0dbfbf8e43cfb0a09782eaeb5ab799fa71