[bazel] Add initial BUILD files for more mojo kernels (#37727)
These are super basic, and right now deps aren't handled. We could add
them but it doesn't affect if the mojopkg is buildable or not. Once we
start making these deps of tools we'll have to support that. But this
works now for lit tests.

MODULAR_ORIG_COMMIT_REV_ID: 9ef168a942d274ac1737f27cc570a784e10df6e8