[LLCL] [GraphRT] Drop cancelation support (#36438)
This appears to be a lookaside value that is stored in the runtime, that
is used only by GraphRT for cancelation. However, it does not have any
usage outside of tests. This is likely a feature that was added early
and did not find purchase. This change proposes dropping for simplicity,
as the cancelation tests in every operation incur a small cost w/o clear
benefit. Especially as the runtime becomes more of a shared resource for
sophisticated pipelines (across Mojo and the engine), the semantics
become less clear.

MODULAR_ORIG_COMMIT_REV_ID: 64a540b03b9de06d4e7da2789f1b64d8ee7f23df