[KGEN] Fix index parameter verifier hopefully for good (#37831)
An index parameter reference can refer to a parameter that exists in a
different scope. I.e. the reference and the declaration are in different
scopes. That means the types of the reference and declaration are in
different scopes. To correctly adjust index references within the type
to be at the correct scope for comparison in the verifier, we know the
declaration is going to be M scopes up, where M is the depth of the
reference. Use that, instead of `signatures.size()-1`, as the adjust
depth.

Fixes [Internal link]

MODULAR_ORIG_COMMIT_REV_ID: 1c48f2463e7b6fd4e2bf265583ed4baf15ded7f2