[Support] Enable non-interactive `modular auth` flow. (#37688)
Currently, `modular auth` only supports an interactive flow that follows
the OAuth Device Authorization spec. It asks the user to visit a URL in
the browser to login and approve the session.

But we need to also support non-interactive flows for:

1. Organizations: To run MAX Engine with usage metering enabled.
2. Users: To build Docker images without user interaction.

This PR refactors the entitlement store and skip the interactive flow if
the `MODULAR_ACCESS_TOKEN` environment variable is set, and use the
access token to diretly to request an X.509 certificate in this case.

MODULAR_ORIG_COMMIT_REV_ID: 96a946a5d789ff7dfeb38eb5b62d7bfbb9346693