[Support] Remove unused telemetry code after Settings migration (#35455)
After the migration to settings and the context, this code is no longer
used.

While this change was originally a clean delete, the underlying code
changed as the stack evolved. As a result, there are two minor related
changes that are included in the delete:
1) There are a few extra uses of `emplaceIfNeeded` which are replaced
with `get`.
2) The tests have been updated to use the settings-based constructors.

MODULAR_ORIG_COMMIT_REV_ID: 46d830a53216988da4ab1f785002197dad01f62c