[mojo-lang] Support declaring functions with both optional and variadic arguments (#37819)
We supported variadics and optional arguments and parameters, but not in
the same signature. This patch fixes that by ensuring that variadics can
have a sentinel default value associated with them that never gets
accessed, but allows us to maintain the invariant of "default correspond
to the trailing arguments". The patch enables both positional and
keyword variadics in the presence of optional arguments. Positional
variadic parameters after optional parameters are also enabled.

Closes [Internal Link]

modular-orig-commit: f37e9334c32fe0d3bf5e751e43886344e3b09bee