[DebugInfo][KGEN] move callLoc out of attr-dict of InlinedSubprogramScoped (#37473)
The callLoc being in an attribute blocks CSE because the location in the
attr-dict is considered part of equivalence between operations. This
patch adds an encoding for callLoc to be in the location of an op
instead.

Closes [Internal link]
MODULAR_ORIG_COMMIT_REV_ID: c699553c340ca25a6047c784e3da1bca5a658b32