[mojo-lang] Rename `ASTDecl::getSelfType`, NFC. (#38274)
This renames this member (and the corresponding setter) to
`ASTDecl::getTypeDeclSelf()`.  The rationale here is that Func
decls also have a self type often and it is confusing to have two
concepts referring to 'self' in different ways.

This also packs some bools in ASTDecl to save space.

MODULAR_ORIG_COMMIT_REV_ID: 5e22b7ddf9695f3383158900514e417aad207c83