[mojo-lang] Implement support for address space param inference. (#38283)
This works around a representational problem with !lit.ref and
!kgen.pointer referring to an unpacked user defined-type instead
of holding the user defined type directly.  It is arguable whether
we want to change the IR (it could make it a lot more noisy and
bloated...) so just add a hack to unblock progress.

MODULAR_ORIG_COMMIT_REV_ID: cf51a5cce71ccd37bbf102e0a80381a78ffb880d