[mojo-stdlib] Remove unconstrained mixed dtype pow overloads. (#38385)
This removes the `x**y` operations that don't constrain the RHS
type (but keeps the one with an Int RHS).  The problem with these
is that they are significant footguns in cases like `x**0.5`
because the compiler infers the RHS to be F64 (the materialization
type of FloatLiteral) regardless of what the LHS type is.

The pow() global function documents that the LHS and RHS types
must be the same, so use the compiler to enforce this.  I discovered
this while working on improving type inference.

MATH_ORIG_REV_ID: 30db26e3fe29a6818cbe6df64ccb0c6f329d0522