[GenericML] Remove the alignment field from `BufferRef` (#38443)
`GML::BufferRef` supports a create/take/borrow, alignment is only needed
when we are creating an owned buffer where the structure is doing the
allocation. In the case where we are taking/borrowing, we do not need to
store the alignment at all. To avoid this weird state where alignment
needs to be tracked for those cases, we should just never store the
alignment.

This simplifies representation of a `GML::BufferRef` and allows easier
conversion to and from `NDBuffer`

MODULAR_ORIG_COMMIT_REV_ID: 5203f1406b890e3042afd39d6c6013187faef103