[mojo-lang] Implement more powerful source location magic (#38087)
We have some ability to get the source location using an existing parser
hack, but this doesn't allow us to get the location of a callsite from
within a function body, which is needed for better asserts. This patch
implements this functionality in the compiler, without using parser
tricks at all. This relies on `@always_inline` or
`@always_inline("nodebug")` decorated functions and some special
handling of the `kgen.source_loc` op in the pre- and post-elaboration
inlining logic. As a consequence, this doesn't immediately work if the
location is within a parameter expression (e.g. inside alias
expressions). The `InterpreterState` API is extended, allowing
interpreter hooks to inspect the call stack.

MODULAR_ORIG_COMMIT_REV_ID: 46826217d0031fb6d70afdcb651cad45bf76f730