[mojo-lang][CompileTime] Cache implicit convertibility, NFC. (#38528)
This teaches the implicit convertibility check to cache its result.
This predicate is both slow (it has to do parameter inferences,
type checking, trait table checking etc etc) and commonly needed
(called in many places in type checking and overload resolution).

Instead of computing it over and over again, cache the results in
SharedState.  This speeds up parse time of expressions.mojo from
6.24 to 5.55s and build stdlib build from 17.21 to 15.24s on
my laptop with a debug build of the compiler.  It is probably
a bigger win on more complex parametric code that is more
dependent on parameter inference.

This is a bit dangerous - I believe (and have empirically verified
on stdlib and mojo testsuite) that it is enough to just include the
two types as the cache key, but it is possible that future changes
(eg introduction of extensions) could make this check context
sensitive. As such, out of paranoia, I have the compiler self check
itself every 64th hit to make sure that we'll find out if things
start diverging.

MODULAR_ORIG_COMMIT_REV_ID: 450a5468593f7d082149430a5ba9525cd8f2317c