[Profiling] Use unordered_set for StringArena (#38486)
Use an `unordered_set` as `StringArena` instead of the previous
`BlockList<std::string>`.
This is because the majority of profiler events' names are static
strings
and the ratio of events to names is large.

NOTE: `Label::intern()` returns a pointer into this `StringArena`.
So it is important that those pointers are still valid after further
insertion into the `StringArena`.
`unordered_set` satisfies this property, but note that for example
`llvm::DenseSet` does not.

When profiling float32 Mistral 7B PyTorch through `mt` at compile time,
this reduces about 60GB of string allocations to an amount negligible
compared to the model size.

MODULAR_ORIG_COMMIT_REV_ID: 3c4312e8e32e6526d5cdc884c8cffd5da68503eb