[Kernels] consolidate accum_type logic to helper methods (#38514)
We have a lot of duplicated code to determine the right type to use for
accumulation operations, and require in a few cases that the caller make
this determination. This change sets a reasonable default but still lets
the caller override the default if necessary.

GPU_TEST_ORIG_REV_ID: 6cb7d8605e08417cb108bcf4ff8ebdbc6182d02e