[mojo-parser] Optimize lookupAndResolveMangledDecl for bytecode containers (#38595)
If the container was loaded from bytecode, we don't need to go through the
normal/complex lookup process, it's guaranteed to already have the right
metadata and be in the container decl. This removes the resolution of a lot
of unnecessary decls within the bytecode module.

This dropped parse time of `print("hello world")` by a third.

MODULAR_ORIG_COMMIT_REV_ID: 380e8cb7db1c8cc30dd551e90c6d52fdff89d000