[mojo-tooling] Restructure how tests are passed to the executor (#38731)
The current structure of the executor involves `mojo` sending just the
test id, and mojo-test-executor reparsing the IR for the test file
and performing checks/extraction/etc to pull out the tests. This is extremely
inefficent and also unnecessary, because `mojo` already performs verification
that the test exists when it resolves tests before execution. This PR takes
advantage of that fact and reworks the executor to instead take in a set of
tests to execute. We can use the information from the test to pull in the tests:
for unit tests we already don't need to parse anything (we just import the
parent module and run the test). For doc tests, the test contains the range of
the code block in the source file (which we can use to extract the code block
and run it).

This is significantly more efficient and will simplify the amount of caching we
need to add for the test infrastructure now that test resolution in the executor
is now effectively instant.

MODULAR_ORIG_COMMIT_REV_ID: dfc54cc3f330fa734d4745e1794bb955e87436c6