[mojo-lang] Use parameter inference for overload set resolution (#38818)
This changes overload resolution for method lookup to not implicitly
bind the parameters of `self` before doing overload resolution.  This
forces the type inference logic to infer the parameters on `Self` from
the actual value provided by `self`.

This new approach allows us (but does not switch us) to have methods
in the overload set with different types for the first argument (the
self
argument).

This is necessary for a few things:
1) We currently have hacks for Non-materializable types, where
    methods lookup looks to see whether the actual type has a
    method and then falls back to the materialization type if not.
    Throwing all the methods into one overload set allows more
    flexible disambiguation.
2) We need to allow self to be declared as `Reference` instead of
    forcing low-level `!lit.ref` usage for parametric mutability.
3) We want `x.foo` where x is a `Reference[T]` to resolve to
     either `Reference.foo` or `T.foo` when autoderef happens.

All of these usecases end up with a method that can have unusual
parameter inference and eventual parameter sets that are not just
`ParamBindings::getForDeclaredType(callersSelf)`.

This is quite a stress test for parameter inference, which has been
the source of the shaking I've been doing on this for the last week.

MODULAR_ORIG_COMMIT_REV_ID: fc2de26a32402bc8fad4b562067554676a9c9449