[mojo-parser] Revert movement of package dep resolution (#38842)
Packages immediately get resolved to signature when they are loaded,
which means that after this was moved in #38508, we weren't loading
in deps anymore. This PR reverts this move to unbreak the serving package,
and gives time to followup later with a test case to catch this in the future.

MODULAR_ORIG_COMMIT_REV_ID: 02ef2cfbb838f69864d87c680da9055216032ec9