[External] [stdlib] Add `repr()` function and `Representable` trait (#38803)
[External] [stdlib] Add `repr()` function and `Representable` trait

This PR is a follow up on the discussion there:

https://github.com/modularml/mojo/issues/2207

There is no `RepresentableRaising` trait yet, this can come later on.

We assume here that `42` is the Mojo representation of `Int(42)` because
this is how it materializes. Same we assume that `"hello"` is the Mojo
string representation of `String("hello")` because StringLiteral
materializes into String.

We could have used `Int(42)` and `String("hello")` as Mojo
representations but that seems very cumbersome and not aligned with the
evolution of the language. The discussion is still open for other types.

The `__repr__` implementation for String is very simplistic and I
believe it's enough as a start, we can push the complexity in other
separate PRs (handling `\` and so on). I will open an issue when this PR
is merged requesting contributor help to improve it.

I will create another issue after this PR is merged to track the
conformance of (all?) library types to the `Representable` trait.

Co-authored-by: gabrieldemarmiesse <gabrieldemarmiesse@gmail.com>
Closes modularml/mojo#2361
MODULAR_ORIG_COMMIT_REV_ID: 82c5e2b83582a61341b8e4a5bfa9d66f7c4f34ab