[bazel] Add name to lit_tests macro (#38918)
This allows us to create a test_suite with a sensible name that runs all
the underlying tests, which allows you to run `./bazelw test
Support/test` vs having to use `:all` or `/...`. It also makes it easier
to run the tests in the directory in the case the tests are marked as
manual and therefore doing `:all` does nothing. This also allows us to
have multiple lit_tests macros in a single BUILD file if it makes sense
to separate suites for different sizes.

MODULAR_ORIG_COMMIT_REV_ID: 75d0e86ca486869787cd295e5bc92fad94e2f26b