[DebugInfo] add decorator info to SourceName attr (#38541)
When a compiler decorator resolves to a function reference, the source
name of the function is now added to the SourceName attribute of
debuginfo. In the long run, all compiler decorators (IE not the
signature and body decorators which are built in to the compiler)
“should” be functions or closures that can do some kind of IR
transformation. At the moment, they can be an arbitrary expression that
evaluates to anything, because there is no verification of what compiler
decorators are.

Results in dwarf that looks like this (from dwarfdump):

```
 DW_AT_name      ("pkg stdlib::pkg collections::module list::fn _max(pkg stdlib::pkg builtin::module int::struct Int @[pkg stdlib::pkg utils::module _visualizers::fn lldb_formatter_wrapping_type],pkg stdlib::pkg builtin::module int::struct Int @[pkg stdlib::pkg utils::module _visualizers::fn lldb_formatter_wrapping_type])")
```

Closes [Internal link]
MODULAR_ORIG_COMMIT_REV_ID: 663d1440a1aa87b1e95eb586807a5acd083a8635