[mojo][debuginfo] Handle pushing potentially null debug scope for functions (#39025)
Now that it's possible for functions to not have debug scope, update the
checks everywhere in the parser so that the null scope can just be
pushed too. NFC if no additional lexical blocks are created in the
function, which is the case for a lot cases.

MODULAR_ORIG_COMMIT_REV_ID: 8dd5967595ad1664dfa91958c2381b77a0922d59