[KGEN] Implement store on `StructGERAttr` (#39010)
Implementing a field-sensitive store is a bit complicated. First, change
the allocation ops to populate a default undef value into memory. Make
loads of StructGERAttr of an undef base value also return undef.

In the store interpret hook, it has to reconstruct the full field access
from the base value, and rebuild the values starting from the bottom
element. If the memory is uninitialized at any point, it needs to lookup
the type definition in order to populate the right LITStructAttr. This
is done via a new hook in the interpreter.

MODULAR_ORIG_COMMIT_REV_ID: 4cb78170e410558deda7db82dce2200b397d9615