[Kernels][GPU] Disable failable compilation by default (#39114)
We want to make sure that compilation failures are not silently ignored
in the 99% of cases. The only cases where we want to ignore them is in
special cases. This adds an option to compile_info to check if the
compilation is failable or not.

COMPILE_TESTS_ORIG_REV_ID: 4a0b019d78083ab44406a4492da1048cb220e7bd