[KGEN][MOGG] Annotate kernels with arg info and remove preserved attr (#39082)
Annotates kernels at the KGEN level with new targeted annotations which
replace the preserved attribute on all generators. This simplifies the
lowering code and also reduces the amount of IR we need to carry around.
We just capture the parameter values as seen on the lit types and the
name of the lit type. The parameters are then preserved through
LowerLitTypes while preserving the order which is exactly what we can't
get from KGEN alone.

MODULAR_ORIG_COMMIT_REV_ID: d1d45e47297269571a02f2a583643e09e63c63a1