[mojo-lang] Move generator of constructor result slot into `emitCallUnchecked` (#39228)
Constructors can be invoked directly like `x.__init__()` or
`Foo.__init__(x)`, but they are often invoked in the form `x = Foo()`.
The more common form is actually a bit of Python magic that Mojo
implements via `emitConstructorCall`. The problem is that with `inout
self` constructors, this function must always create an MLValue slot,
even if the function call folds to a constant, in which case it can
return a PValue, instead of an MLValue.

This PR moves the logic for generating an MLValue slot for constructor
calls into `emitCallUnchecked`. This piggy-backs off the similar logic
for `byref_result` slots. The function distinguishes between the magic
form and the direct call form by checking if a non-null value has been
provided for the `init_self` argument.

MODULAR_ORIG_COMMIT_REV_ID: 2fe1b0800c03460feee6614301694de686cd79ec