[KGEN] Drop the result types on `!co.routine` (#39325)
This is a pretty big change to the async function type system. This PR
drops the list of result types on the `!co.routine` type, turning it
into a single, opaque type. We instead rely on the `Coroutine` type in
the library to track the correct result types through the type system.

This reduces the verbosity of the IR and moves towards a simpler IR
model. At the same time, it removes the need to use `pointer<i8>` as a
hack to represent opaque coroutines.

At the same time, remove the `__call__` method on `Coroutine`. Directly
resuming a coroutine is something that has to be done explicitly. Making
it easily accessible via `()` is a huge trap, because of the async
function suspends, the caller will not be ready to handle it.

RUNTIME_ORIG_REV_ID: c103e5ca6e3b645fe73f468497a88be73811d439