[mojo-lang] Introduce a new `TypeCheckScopeInfo` struct, NFC (#39361)
This gives a name to the declScope+shared pair that is passed
pervasively around the parser, and adds a new "isParamContext"
bit that indicates whether the type checking is happening in
a parameter expression context.  The former is a cleanup, the
later enables us to know whether we are evaluating a parameter
expression for future use.

MODULAR_ORIG_COMMIT_REV_ID: 9ffd1f1ba6d24e473c70a0c91a6b25e74f56ac6f