[mojo-lang] Allow binding `!lit.ref` args in parameter expression. (#39364)
One big limitation of parameter expression has been that you can't
bind a Reference (or similar things that use parameteric mutability)
in a parameter expression, because you just have a PValue not an
MValue for the argument.  This is a big problem for Optional and
Reference and other similar types in parameter expressions.

This patch enhances the typechecker to know that it is fine to do
this in a parameter context, because we can drop the value into an
immortal interpreter memslot and infer from that.

This fixes MOCO-584 - [References] We cannot bind !lit.ref in parameter
context

MODULAR_ORIG_COMMIT_REV_ID: 4ba0b848efa3651b917e5f89f22cafc17fb3ec71