[mojo-lang] Enable use of `Reference` for self arguments. (#39367)
Finally after a ton of grooming we're able to enable the direct use
of `Reference` as a 'self' argument of a method instead of requiring
everyone to use the raw `!lit.ref` type.  This required fixing the
overloadsets to not hardcode the parameter, required being able to
infer reference bindings in parameter expr contexts and lots of
other things.

MODULAR_ORIG_COMMIT_REV_ID: 19eb45170f8326ea2ae4d364be7eb858145905f7