[Engine API] Add session-wide compile configuration. (#39306)
This PR adds a session-wide (`globalCompileConfig`) compile
configuration that is shared among all models instantiated from that
session.

The motivating use-case for this is sharing the same custom ops library
among all users of the session; using Mojo custom operators in PyTorch
eager and `torch.compile` will use this in upcoming PRs. This is
essential because the custom operations need to be registered before a
model is compiled or loaded.

MAX_PYTHON_ORIG_REV_ID: 82e9e52454ef18fd8d0124c900a7048e52396807