[mojo-tooling] Add support for erroring out on invalid doc strings
We currently emit warnings for invalid aspects of doc strings, but this
makes it difficult to actually enforce that all doc strings are valid. The
current strategy involves resorting to inspecting the stderr of a run
of `mojo doc` to see if any warnings were emitted, which is not ideal.
This PR restructures the processing of doc strings in two ways to make
this easier:

* A new `validate-doc-strings` flag is added, that will turn doc string
related diagnostics into errors (instead of warnings).
* The `warn-missing-doc-strings` flag is renamed to
`diagnose-missing-doc-strings` to better reflect its purpose (and not
attach too strongly to "warnings").

MODULAR_ORIG_COMMIT_REV_ID: 4f2f8c1413ff9659d1eb95ea317c46024e666837