[mojo-lang] Error Branch Does Not Always Not Consume EndInitMaybe + CheckDef for Conditional Calls (#39075)
If a conditionally initialized value is initialized in the error branch (the
case when an exception is caught and not re-thrown), check lifetimes would trigger a destruction in
the success branch. Separately, we are missing a check for destruction before the overwrite
in the conditionally initializing call. This patch addresses both of those issues.

MODULAR_ORIG_COMMIT_REV_ID: 39a5f21dd6a088bb4fddf451748270d42ce0504b