[DebugInfo] Restore xfailed debug-jit test (#39318)
Part 1 of [MOTO-429 : Changing stdlib `print` to be variadic brings
debug-info pipeline for lifetime tracking of printed
[Internal link]
(Part 2 requires actual fix).

There's no need for the column number in this test case as it's not
important. Previously it was breaking on the call itself (the open
parenthesis) because that's the first instruction of that line, but now
it's breaking on the first argument (the opening quotation mark) because
it needs to be packaged into a variadic pack first.

MODULAR_ORIG_COMMIT_REV_ID: 27ba51c60a0d7b7200b9dfaec8ab36c9754c9697