[Support][DebugInfo][GPU] Add NVPTX conversion for LineTableLocOp (#39423)
NVPTX does not have a "nop" instruction. Add a custom lowering for the
LineTableLocOp into a nop-equivalent instruction. Currently using
"[pmevent.mask
0](https://docs.nvidia.com/cuda/parallel-thread-execution/index.html#miscellaneous-instructions-pmevent)"
(which is supposed to do nothing) and it's trivial to change this later.
I've tested to make sure it is preserved by the nvptx compiler (such a
test will be added with the PR that inserts LineTableLocOps).

MODULAR_ORIG_COMMIT_REV_ID: 0bad08acba494b52c9e4634b144da7bbfefec8b1