[KGEN] Stop checking `ParameterAttr::isConstant` in IREvaluator (#39500)
The model KGEN has of whether an attribute is really a constant when it
comes to parameter expressions is completely broken and is only
beneficial as a compile time optimization at this point. Stop checking
it in the elaborator because it's not a reliable indicator if things are
correct. In the elaborator, we have the benefit that things *must*
concretize or error, so we can rely on clients to check.

This fixes the issue @lattner ran into in
[Internal link]

MODULAR_ORIG_COMMIT_REV_ID: ba28aef7332a4ef2e69b6ad935f7fe96444373f9