[LSP] Detect unused function arguments (#39073)
This PR detects unused function arguments as an addon to the previous PR
in this stack. Right now, we flag `self` as unused in struct methods;
this may or may not be desirable in the long run, but avoiding it takes
a fair bit of work.

MODULAR_ORIG_COMMIT_REV_ID: bd79a7bf0158fe46bc3f215d89aacd69172e9e80