[MGP][GenericML] Switched to a static tensor spec for primitives (#39679)
When we are generating functions for our mgp primitives, to conform to
DPS, we insert an explicit pop.stack_allocation to reserve space for the
creation of a TensorSpec struct.  This struct will heap-allocate in it's
ctor which is called during `mgp_tensor_spec_create`  See snippet here:
```
kgen.generator export @__stub_6b84071145db0836(%arg0: !kgen.pointer<pointer<scalar<invalid>>>, %arg1: !kgen.pointer<scalar<invalid>>, %arg2: !kgen.pointer<none>) {
  kgen.param.declare simd_width = <1>
  %index0 = kgen.param.constant = <0>
  %0 = pop.offset %arg0[%index0] : !kgen.pointer<pointer<scalar<invalid>>>
  %1 = pop.load %0 align<8> : !kgen.pointer<pointer<scalar<invalid>>>
  %2 = pop.variadic.create [] : !kgen.variadic<none>
  %3 = kgen.rebind %2 : !kgen.variadic<none> to !kgen.variadic<index>
  %4 = pop.stack_allocation 1 x struct<(struct<(struct<(pointer<none>, scalar<si64>)>) memoryOnly>) memoryOnly> align 64
  %5 = kgen.call @mgp_tensor_spec_create<:scalar<ui8> 134, :variadic<variant<index, i1>> [apply(:() -> !kgen.variant<index, i1> @create_known_dim<1>), apply(:() -> !kgen.variant<index, i1> @create_known_dim<2>), apply(:() -> !kgen.variant<index, i1> @create_known_dim<2>), apply(:() -> !kgen.variant<index, i1> @create_known_dim<1>), apply(:() -> !kgen.variant<index, i1> @create_known_dim<1>), apply(:() -> !kgen.variant<index, i1> @create_known_dim<1>), apply(:() -> !kgen.variant<index, i1> @create_known_dim<2>), apply(:() -> !kgen.variant<index, i1> @create_known_dim<2>)], 8>(%3, %4) : (!kgen.variadic<index> borrow, !kgen.pointer<struct<(struct<(struct<(pointer<none>, scalar<si64>)>) memoryOnly>) memoryOnly>> byref_result) -> !kgen.none
  %6 = kgen.call @create_tensor_spec_async<8>(%4, %1, %arg1) : (!kgen.pointer<struct<(struct<(struct<(pointer<none>, scalar<si64>)>) memoryOnly>) memoryOnly>> borrow_in_mem, !kgen.pointer<scalar<invalid>> borrow, !kgen.pointer<scalar<invalid>> borrow) -> !kgen.none
  kgen.return
}
```

I expect the `dtor` of the struct to be called before the function
returns, (which would result in the correct free being called). This is
not the case however.

An alternative is to implement a pass which explicitly inserts a
destructor. But this would require knowledge of the inserted
mgp.tensor_spec.create etc.

However, at the primitives, we already know the rank at compile time,
and in production, we use NDBuffer (which is a staticIntTuple) we want
to switch to a "static" TensorSpec which has a known rank. As such, we
implemented a `StaticTensorSpec` which wraps the `StaticIntTuple` and
`DType`

Fixes GRA-454

MODULAR_ORIG_COMMIT_REV_ID: 726847b3eed0ef4893d9b64f75ad614ed5c7d07f