[Engine API] Set MOJO_PYTHON_LIBRARY in API (#38992)
When using the Engine API to run a model with custom ops that are
written in Mojo but use our Python interop, Mojo will be unable to find
the Python library since we did not go through the `mojo` tool. Deal
with this in the Engine API as well.

In this case we already have a configuration read inside `Settings`, and
we would not want to read it again for this, so some work was needed to
get `MojoConfig` to work with `Settings`.

Fixes MSDK-97.

---------

Co-authored-by: Jakub Tucholski <jakub@modular.com>
MODULAR_ORIG_COMMIT_REV_ID: 83c7e1be8116475e7cf61fa0937af598aac3d319