[Debug] Improve target launch in tests (#39699)
I was using LaunchSimple for launching the target in our tests, but
that's not a good API because it hides any errors that happen. Instead,
I'm now using a different API that returns the error that we can show to
the user. That would have been helpful to Keith when debugging an issue
in these tests.

I also took the chance to change an erroneous assert for a
report_fatal_error.

MODULAR_ORIG_COMMIT_REV_ID: bafc200491e35c5142fe2f8f9d10099b3ed480a9