[mojo-vscode] Restructure how SDKs are detected (#39727)
This PR performs some fundamental restructuring around how
the extension handles Mojo SDKs, both with detection and the
tracking.

The first major change is that the extension will now only use a single
SDK, and no longer try to manage multiple SDKs for different
workpsace folders. There isn't a real benefit to having this kind of
optionality, and it leads to very complicated tracking and management
(accounting for all of the possible SDK usage situations). Moving forward
after an SDK is resolved, that SDK is used for all files.

The second major change is that the extension will now detect SDKs that
are installed via the modular cli tool. This is a significant change for
external users as they should no longer ever have to set `modularHomePath`
when installing the SDK, we should pick it up directly from the cli.
`modularHomePath` still exists, and can be set, but that should soon be
relegated to just internal use cases.

MODULAR_ORIG_COMMIT_REV_ID: dab327368ace1043c2ded2214eda50ceb5717b6a